Change Line Ending, Cache SharedString File
XLSX.js' line endings should now be LF, instead of CRLF. That change
makes every line show as changed, even though the content of most lines
have not. The one other change is on lines 24 and 25. Instead of calling
zip.file('xl/sharedStrings.xml').asText() twice (once in evaluating the
condition, and once when true), the file part should be cached separate
for performance and stability. If the file exists, it would otherwise do
the UTF8 encoding twice when the first use is thrown away. If the file
doesn't exist, it would error because there is no asText property on
null.