Read blank cells, maxCol, maxRow, write formulas
When reading a document, blank cells (including blank rows) are no
longer eliminated in the resulting JS object. Formulas (stored in a
cell's "formula" property) are now written to document. maxCol and
maxRow are now a property of worksheets, when read, describing the max
dimensions of the array(s). Also, misc. cleanup.